requirements.txt
rich>=13.0
requests>=2.28
python-whois>=0.8
phonenumbers>=8.13
qrcode>=7.4
dnspython>=2.3
Faker>=18.0
pyfiglet>=1.0
psutil>=5.9
paramiko>=3.0
Pillow>=10.0
scapy>=2.5
pynput>=1.7
